Добавляет функцию вызова объявлений
